refact files in Services (#70)
* refact files in Services

* refactor services

+++ services/organizations.js
@@ -1,27 +1,57 @@
 const organizationsRepository = require('../repositories/organizations');
 
-async function getOrganizationPublic(id) {
-  return await organizationsRepository.getOrganizationPublic(id);
-}
+const getOrganizationPublic = async (id) => {
+  const organization = await organizationsRepository.getOrganizationPublic(id);
+  if (!organization) {
+    const error = new Error(`No existe la organización`);
+    error.status = 404;
+    throw error;
+  }
+  return organization;
+};
 
-async function getAll() {
-  return await organizationsRepository.getAll();
-}
-async function getById(id) {
-  return await organizationsRepository.getById(id);
-}
+const getAll = async () => {
+  const organizations = await organizationsRepository.getAll();
+  if (organizations.length == 0) {
+    const error = new Error(`No existen organizaciones`);
+    error.status = 404;
+    throw error;
+  }
+  return organizations;
+};
+const getById = async (id) => {
+  const organization = await organizationsRepository.getById(id);
+  if (!organization) {
+    const error = new Error(`No existe la organización`);
+    error.status = 404;
+    throw error;
+  }
+  return organization;
+};
 
-async function create(organization) {
+const create = async (organization) => {
   return await organizationsRepository.create(organization);
-}
+};
 
-async function update(id, organization) {
+const update = async (id, organization) => {
+  const response = await organizationsRepository.getById(id);
+  if (!response) {
+    const error = new Error('La organización no existe.');
+    error.status = 404;
+    throw error;
+  }
   return await organizationsRepository.update(id, organization);
-}
+};
 
-async function remove(id) {
+const remove = async (id) => {
+  const response = await organizationsRepository.getById(id);
+  if (!response) {
+    const error = new Error('La organización no existe.');
+    error.status = 404;
+    throw error;
+  }
   return await organizationsRepository.remove(id);
-}
+};
 
 module.exports = {
   getAll,
@@ -29,5 +59,5 @@ module.exports = {
   create,
   update,
   remove,
-  getOrganizationPublic,
+  getOrganizationPublic
 };
